Fixed map and added more stuff

+++ src/Player.cpp
@@ -130,7 +130,7 @@ namespace v2 {
 	void Player::loadResources() { return (this->*union_cast<void(Player::*)()>(0x46c0b0))(); }
 	bool Player::updateGroundMovement(float value) { return (this->*union_cast<bool(Player::*)(float)>(0x487740))(value); }
 	bool Player::updateAirMovement(float a0, float a1) { return (this->*union_cast<bool(Player::*)(float,float)>(0x4877C0))(a0, a1); }
-	void Player::addCardMeter(int value) { return (this->*union_cast<void(Player::*)(float)>(0x487870))(value); }
+	void Player::addCardMeter(int value) { return (this->*union_cast<void(Player::*)(int)>(0x487870))(value); }
 	void Player::Unknown487C20() { return (this->*union_cast<void(Player::*)()>(0x487C20))(); }
 	void Player::playSpellBackground(int a0, int a1) { return (this->*union_cast<void(Player::*)(int, int)>(0x46b370))(a0, a1); }
 	bool Player::applyGroundMechanics() { return (this->*union_cast<bool(Player::*)()>(0x487ca0))(); }
